Save regarded parameters and select new config based on that
First, generate a order in which to choose the parameter values
and when to check the different constraints based on that.

For every configuration generation, first select a random value
for the first parameter, check if it fulfills possible constraints
and isn't completely explored yet. If this takes too many tries,
list all valid values for the parameter and select one of them.
Continue recursively. If no valid value can be found for a specific
parameter, go back one parameter and find a new value there.